Week 9/requirements.txt
pygame-ce==2.5.6
fpdf2==2.7.9
pytest==9.0.1